removed bus_types, rename generic SerialPort to Port to avoid overlap
