+++ src/bus/mod.rs
@@ -1,7 +1,45 @@
 //! Low level interface to a DYNAMIXEL Protocol 2.0 bus.
 
-use crate::bus_types::{bytestuff, endian, InstructionPacket, Packet, StatusPacket, HEADER_PREFIX, HEADER_SIZE};
-use crate::{checksum, ReadError, WriteError};
+#[path = "."]
+pub(crate) mod asynch {
+	use crate::asynch::SerialPort;
+	use bisync::asynchronous::*;
+	mod bus;
+	pub(crate) use bus::Bus;
+}
+#[path = "."]
+pub(crate) mod sync {
+	use crate::SerialPort;
+	use bisync::synchronous::*;
+	mod bus;
+	pub(crate) use bus::Bus;
+}
+
+// pub(crate) use asynch::bus::Bus as AsyncBus;
+// pub(crate) use sync::bus::Bus;
+
+use core::time::Duration;
+
+pub(crate) mod bytestuff;
+pub(crate) mod endian;
+
+pub(crate) mod data;
+pub use data::Data;
+
+mod packet;
+pub use packet::{InstructionPacket, Packet, StatusPacket};
+
+/// Prefix of a packet.
+///
+/// All packets start with this prefix, and they can not contain it in the body.
+///
+/// In other words: if you see this in the data stream, it must be the start of a packet.
+pub(crate) const HEADER_PREFIX: [u8; 4] = [0xFF, 0xFF, 0xFD, 0x00];
+
+/// The size of a message header, including the pre-amble, packet ID and length.
+///
+/// Excludes the instruction ID, the error field of status packets, the parameters and the CRC.
+pub(crate) const HEADER_SIZE: usize = 7;
 
 /// Default buffer type.
 ///
@@ -17,250 +55,40 @@ pub type DefaultBuffer = alloc::vec::Vec<u8>;
 #[cfg(not(feature = "alloc"))]
 pub type DefaultBuffer = &'static mut [u8];
 
-/// Low level interface to a DYNAMIXEL Protocol 2.0 bus.
+/// Create a mutable static buffer of size N.
 ///
-/// Does not assume anything about the direction of communication.
-/// Used by [`crate::Client`] and [`crate::Device`].
-pub(crate) struct Bus<SerialPort, Buffer>
-where
-	SerialPort: super::SerialPort,
-	Buffer: AsRef<[u8]> + AsMut<[u8]>,
-{
-	/// The underlying stream (normally a serial port).
-	pub(crate) serial_port: SerialPort,
-
-	/// The baud rate of the serial port, if known.
-	pub(crate) baud_rate: u32,
-
-	/// The buffer for reading incoming messages.
-	pub(crate) read_buffer: Buffer,
-
-	/// The total number of valid bytes in the read buffer.
-	pub(crate) read_len: usize,
-
-	/// The number of leading bytes in the read buffer that have already been used.
-	pub(crate) used_bytes: usize,
-
-	/// The buffer for outgoing messages.
-	pub(crate) write_buffer: Buffer,
-}
-
-#[super::bisync]
-impl<SerialPort, Buffer> Bus<SerialPort, Buffer>
-where
-	SerialPort: super::SerialPort,
-	Buffer: AsRef<[u8]> + AsMut<[u8]>,
-{
-	/// Create a new bus using pre-allocated buffers.
-	///
-	/// The serial port must already be configured in raw mode with the correct baud rate,
-	/// character size (8), parity (disabled) and stop bits (1).
-	pub fn with_buffers(serial_port: SerialPort, read_buffer: Buffer, write_buffer: Buffer) -> Result<Self, SerialPort::Error> {
-		let baud_rate = serial_port.baud_rate()?;
-		Ok(Self::with_buffers_and_baud_rate(serial_port, read_buffer, write_buffer, baud_rate))
-	}
-
-	/// Create a new bus using pre-allocated buffers.
-	pub fn with_buffers_and_baud_rate(serial_port: SerialPort, read_buffer: Buffer, write_buffer: Buffer, baud_rate: u32) -> Self {
-		let mut write_buffer = write_buffer;
-
-		// Pre-fill write buffer with the header prefix.
-		// TODO: return Err instead of panicking.
-		assert!(write_buffer.as_mut().len() >= HEADER_SIZE + 3);
-		write_buffer.as_mut()[..4].copy_from_slice(&HEADER_PREFIX);
-
-		Self {
-			serial_port,
-			baud_rate,
-			read_buffer,
-			read_len: 0,
-			used_bytes: 0,
-			write_buffer,
+/// This macro returns a `&'mut static [u8]`.
+/// Each occurence of the macro may only be evaluated once,
+/// any subsequent invocation will panic.
+///
+/// The macro is usable as expression in const context.
+///
+/// # Usage:
+/// ```no_run
+/// # fn main() -> Result<(), std::io::Error> {
+/// # let serial_port = serial2::SerialPort::open("/dev/null", 57600)?;
+/// use dynamixel2::{Client, static_buffer};
+/// let client = Client::with_buffers(serial_port, static_buffer!(128), static_buffer!(64))?;
+/// # Ok(())
+/// # }
+/// ```
+#[macro_export]
+macro_rules! static_buffer {
+	($N:literal) => {{
+		use ::core::sync::atomic::{AtomicBool, Ordering};
+		static USED: AtomicBool = AtomicBool::new(false);
+		static mut BUFFER: [u8; $N] = [0; $N];
+		if USED.swap(true, Ordering::Relaxed) {
+			panic!("static buffer already used, each occurence of `static_buffer!()` may only be evaluated once");
 		}
-	}
-
-	/// Set the baud rate of the underlying serial port.
-	pub fn set_baud_rate(&mut self, baud_rate: u32) -> Result<(), SerialPort::Error> {
-		self.serial_port.set_baud_rate(baud_rate)?;
-		self.baud_rate = baud_rate;
-		Ok(())
-	}
-
-	/// Write a status message to the bus.
-	pub async fn write_status<F>(
-		&mut self,
-		packet_id: u8,
-		error: u8,
-		parameter_count: usize,
-		encode_parameters: F,
-	) -> Result<(), WriteError<SerialPort::Error>>
-	where
-		F: FnOnce(&mut [u8]) -> Result<(), crate::error::BufferTooSmallError>,
-	{
-		crate::error::BufferTooSmallError::check(StatusPacket::message_len(parameter_count), self.write_buffer.as_ref().len())?;
-		self.write_packet(packet_id, crate::instruction_id::STATUS, parameter_count + 1, |buffer| {
-			buffer[0] = error;
-			encode_parameters(&mut buffer[1..])
-		})
-		.await
-	}
-
-	/// Write an instruction message to the bus.
-	pub async fn write_instruction<F>(
-		&mut self,
-		packet_id: u8,
-		instruction_id: u8,
-		parameter_count: usize,
-		encode_parameters: F,
-	) -> Result<(), WriteError<SerialPort::Error>>
-	where
-		F: FnOnce(&mut [u8]) -> Result<(), crate::error::BufferTooSmallError>,
-	{
-		self.write_packet(packet_id, instruction_id, parameter_count, encode_parameters)
-			.await
-	}
-
-	/// Write a packet to the bus.
-	pub async fn write_packet<F>(
-		&mut self,
-		packet_id: u8,
-		instruction_id: u8,
-		parameter_count: usize,
-		encode_parameters: F,
-	) -> Result<(), WriteError<SerialPort::Error>>
-	where
-		F: FnOnce(&mut [u8]) -> Result<(), crate::error::BufferTooSmallError>,
-	{
-		let buffer = self.write_buffer.as_mut();
-
-		// Check if the buffer can hold the unstuffed message.
-		crate::error::BufferTooSmallError::check(InstructionPacket::message_len(parameter_count), buffer.len())?;
-
-		// Add the header, with a placeholder for the length field.
-		buffer[4] = packet_id;
-		buffer[5] = 0;
-		buffer[6] = 0;
-		buffer[7] = instruction_id;
-		encode_parameters(&mut buffer[8..][..parameter_count])?;
-
-		// Perform bitstuffing on the body.
-		// The header never needs stuffing.
-		// However, strictly following the spec, the instruction ID might need stuffing.
-		let stuffed_body_len = bytestuff::stuff_inplace(&mut buffer[HEADER_SIZE..], 1 + parameter_count)?;
-
-		endian::write_u16_le(&mut buffer[5..], stuffed_body_len as u16 + 2);
-
-		// Add checksum.
-		let checksum_index = HEADER_SIZE + stuffed_body_len;
-		let checksum = checksum::calculate_checksum(0, &buffer[..checksum_index]);
-		endian::write_u16_le(&mut buffer[checksum_index..], checksum);
-
-		// Throw away old data in the read buffer and the kernel read buffer.
-		// We don't do this when reading a reply, because we might receive multiple replies for one instruction,
-		// and read() can potentially read more than one reply per syscall.
-		self.read_len = 0;
-		self.used_bytes = 0;
-		self.serial_port.discard_input_buffer().map_err(WriteError::DiscardBuffer)?;
-
-		// Send message.
-		let stuffed_message = &buffer[..checksum_index + 2];
-		trace!("sending packet: {:02X?}", stuffed_message);
-		self.serial_port.write_all(stuffed_message).await.map_err(WriteError::Write)?;
-		Ok(())
-	}
-
-	/// Read a raw packet from the bus with the given deadline.
-	pub async fn read_packet_deadline(&mut self, deadline: SerialPort::Instant) -> Result<Packet<'_>, ReadError<SerialPort::Error>> {
-		// Check that the read buffer is large enough to hold atleast a instruction packet with 0 parameters.
-		crate::error::BufferTooSmallError::check(HEADER_SIZE + 3, self.read_buffer.as_mut().len())?;
-
-		let stuffed_message_len = loop {
-			self.remove_garbage();
-
-			// The call to remove_garbage() removes all leading bytes that don't match a packet header.
-			// So if there's enough bytes left, it's a packet header.
-			if self.read_len > HEADER_SIZE {
-				let read_buffer = &self.read_buffer.as_mut()[..self.read_len];
-				let body_len = endian::read_u16_le(&read_buffer[5..]) as usize;
-
-				// Check if the read buffer is large enough for the entire message.
-				crate::error::BufferTooSmallError::check(HEADER_SIZE + body_len, self.read_buffer.as_mut().len()).inspect_err(|_| {
-					self.consume_read_bytes(HEADER_SIZE);
-				})?;
-
-				if self.read_len >= HEADER_SIZE + body_len {
-					break HEADER_SIZE + body_len;
-				}
-			}
-
-			// Try to read more data into the buffer.
-			let new_data = self
-				.serial_port
-				.read(&mut self.read_buffer.as_mut()[self.read_len..], &deadline)
-				.await
-				.map_err(ReadError::Io)?;
-
-			self.read_len += new_data;
-		};
-
-		let buffer = self.read_buffer.as_mut();
-		let parameters_end = stuffed_message_len - 2;
-		trace!("read packet: {:02X?}", &buffer[..parameters_end]);
-
-		let checksum_message = endian::read_u16_le(&buffer[parameters_end..]);
-		let checksum_computed = checksum::calculate_checksum(0, &buffer[..parameters_end]);
-		if checksum_message != checksum_computed {
-			self.consume_read_bytes(stuffed_message_len);
-			return Err(crate::InvalidChecksum {
-				message: checksum_message,
-				computed: checksum_computed,
-			}
-			.into());
+		unsafe {
+			// Use raw pointer to avoid compiler warning.
+			let buffer = &raw mut BUFFER;
+			// Convert to reference in separate expression to avoid clippy warning.
+			let buffer = &mut *buffer;
+			buffer.as_mut_slice()
 		}
-
-		// Mark the whole message as "used_bytes", so that the next call to `remove_garbage()` removes it.
-		self.used_bytes += stuffed_message_len;
-
-		// Remove byte-stuffing from the everything from instruction ID to the parameters.
-		let parameter_count = bytestuff::unstuff_inplace(&mut buffer[HEADER_SIZE..parameters_end]);
-
-		// Wrap the data in a `Packet`.
-		let data = &self.read_buffer.as_ref()[..HEADER_SIZE + parameter_count];
-		let packet = Packet { data };
-
-		// Ensure that status packets have an error field (included in parameter_count here).
-		if packet.instruction_id() == crate::instruction_id::STATUS && parameter_count < 1 {
-			return Err(crate::InvalidMessage::InvalidParameterCount(crate::InvalidParameterCount {
-				actual: 0,
-				expected: crate::ExpectedCount::Min(1),
-			})
-			.into());
-		}
-
-		Ok(packet)
-	}
-
-	/// Remove leading garbage data from the read buffer.
-	fn remove_garbage(&mut self) {
-		let read_buffer = self.read_buffer.as_mut();
-		let garbage_len = find_header(&read_buffer[..self.read_len][self.used_bytes..]);
-		if garbage_len > 0 {
-			debug!("skipping {} bytes of leading garbage.", garbage_len);
-			trace!("skipped garbage: {:02X?}", &read_buffer[..garbage_len]);
-		}
-		self.consume_read_bytes(self.used_bytes + garbage_len);
-		debug_assert_eq!(self.used_bytes, 0);
-	}
-
-	fn consume_read_bytes(&mut self, len: usize) {
-		debug_assert!(len <= self.read_len);
-		self.read_buffer.as_mut().copy_within(len..self.read_len, 0);
-		// Decrease both used_bytes and read_len together.
-		// Some consumed bytes may be garbage instead of used bytes though.
-		// So we use `saturating_sub` for `used_bytes` to cap the result at 0.
-		self.used_bytes = self.used_bytes.saturating_sub(len);
-		self.read_len -= len;
-	}
+	}};
 }
 
 /// Find the potential starting position of a header.
@@ -279,13 +107,109 @@ fn find_header(buffer: &[u8]) -> usize {
 	buffer.len()
 }
 
+/// Calculate the required time to transfer a message of a given size.
+///
+/// The size must include any headers and footers of the message.
+pub(crate) fn message_transfer_time(message_size: u32, baud_rate: u32) -> Duration {
+	let baud_rate = u64::from(baud_rate);
+	let bits = u64::from(message_size) * 10; // each byte is 1 start bit, 8 data bits and 1 stop bit.
+	let secs = bits / baud_rate;
+	let subsec_bits = bits % baud_rate;
+	let nanos = (subsec_bits * 1_000_000_000).div_ceil(baud_rate);
+	Duration::new(secs, nanos as u32)
+}
+
 #[cfg(test)]
 mod test {
-	use core::time::Duration;
-
 	use super::*;
-	use crate::static_buffer;
 	use assert2::assert;
+
+	#[test]
+	fn test_message_transfer_time() {
+		// Try a bunch of values to ensure we dealt with overflow correctly.
+		assert!(message_transfer_time(100, 1_000) == Duration::from_secs(1));
+		assert!(message_transfer_time(1_000, 10_000) == Duration::from_secs(1));
+		assert!(message_transfer_time(1_000, 1_000_000) == Duration::from_millis(10));
+		assert!(message_transfer_time(1_000, 10_000_000) == Duration::from_millis(1));
+		assert!(message_transfer_time(1_000, 100_000_000) == Duration::from_micros(100));
+		assert!(message_transfer_time(1_000, 1_000_000_000) == Duration::from_micros(10));
+		assert!(message_transfer_time(1_000, 2_000_000_000) == Duration::from_micros(5));
+		assert!(message_transfer_time(1_000, 4_000_000_000) == Duration::from_nanos(2500));
+		assert!(message_transfer_time(10_000, 4_000_000_000) == Duration::from_micros(25));
+		assert!(message_transfer_time(1_000_000, 4_000_000_000) == Duration::from_micros(2500));
+		assert!(message_transfer_time(10_000_000, 4_000_000_000) == Duration::from_millis(25));
+		assert!(message_transfer_time(100_000_000, 4_000_000_000) == Duration::from_millis(250));
+		assert!(message_transfer_time(1_000_000_000, 4_000_000_000) == Duration::from_millis(2500));
+		assert!(message_transfer_time(2_000_000_000, 4_000_000_000) == Duration::from_secs(5));
+		assert!(message_transfer_time(4_000_000_000, 4_000_000_000) == Duration::from_secs(10));
+		assert!(message_transfer_time(4_000_000_000, 2_000_000_000) == Duration::from_secs(20));
+		assert!(message_transfer_time(4_000_000_000, 1_000_000_000) == Duration::from_secs(40));
+		assert!(message_transfer_time(4_000_000_000, 100_000_000) == Duration::from_secs(400));
+		assert!(message_transfer_time(4_000_000_000, 10_000_000) == Duration::from_secs(4_000));
+		assert!(message_transfer_time(4_000_000_000, 1_000_000) == Duration::from_secs(40_000));
+		assert!(message_transfer_time(4_000_000_000, 100_000) == Duration::from_secs(400_000));
+		assert!(message_transfer_time(4_000_000_000, 10_000) == Duration::from_secs(4_000_000));
+		assert!(message_transfer_time(4_000_000_000, 1_000) == Duration::from_secs(40_000_000));
+		assert!(message_transfer_time(4_000_000_000, 100) == Duration::from_secs(400_000_000));
+		assert!(message_transfer_time(4_000_000_000, 10) == Duration::from_secs(4_000_000_000));
+		assert!(message_transfer_time(4_000_000_000, 1) == Duration::from_secs(40_000_000_000));
+
+		assert!(message_transfer_time(43, 1) == Duration::from_secs(430));
+		assert!(message_transfer_time(43, 10) == Duration::from_secs(43));
+		assert!(message_transfer_time(43, 2) == Duration::from_secs(215));
+		assert!(message_transfer_time(43, 20) == Duration::from_millis(21_500));
+		assert!(message_transfer_time(43, 200) == Duration::from_millis(2_150));
+		assert!(message_transfer_time(43, 2_000_000) == Duration::from_micros(215));
+		assert!(message_transfer_time(43, 2_000_000_000) == Duration::from_nanos(215));
+		assert!(message_transfer_time(43, 4_000_000_000) == Duration::from_nanos(108)); // rounded up
+		assert!(message_transfer_time(3, 4_000_000_000) == Duration::from_nanos(8)); // rounded up
+		assert!(message_transfer_time(5, 4_000_000_000) == Duration::from_nanos(13)); // rounded up
+
+		let lots = u32::MAX - 1; // Use MAX - 1 because MAX is not cleanly divisible by 2.
+		assert!(message_transfer_time(lots, 1) == Duration::from_secs(u64::from(lots) * 10));
+		assert!(message_transfer_time(lots, lots) == Duration::from_secs(10));
+		assert!(message_transfer_time(lots / 2, lots) == Duration::from_secs(5));
+		assert!(message_transfer_time(lots, lots / 2) == Duration::from_secs(20));
+	}
+
+	#[test]
+	#[allow(clippy::needless_range_loop, reason = "indexing gives better error messsage on failed assertion")]
+	fn test_static_buffer() {
+		let buffer1 = static_buffer!(128);
+		assert!(buffer1.len() == 128);
+		for i in 0..buffer1.len() {
+			assert!(buffer1[i] == 0, "i: {i}");
+		}
+
+		let buffer2 = static_buffer!(64);
+		assert!(buffer2.len() == 64);
+		for i in 0..buffer2.len() {
+			assert!(buffer2[i] == 0, "i: {i}");
+		}
+	}
+
+	#[test]
+	#[should_panic]
+	fn test_static_buffer_panics_when_evaluated_in_loop() {
+		for _ in 0..2 {
+			let buffer = static_buffer!(128);
+			assert!(buffer.len() == 128);
+		}
+	}
+
+	#[test]
+	#[should_panic]
+	fn test_static_buffer_panics_when_evaluated_in_loop_through_function() {
+		fn make_buffer() -> &'static mut [u8] {
+			static_buffer!(128)
+		}
+
+		for _ in 0..2 {
+			let buffer = make_buffer();
+			assert!(buffer.len() == 128);
+		}
+	}
+	use core::time::Duration;
 
 	#[test]
 	fn test_find_garbage_end() {
@@ -347,7 +271,7 @@ mod test {
 				endian::write_u16_le(&mut _buffer[offset + 5..], packet_2_length as u16);
 				let checksum_len = 2;
 				let checksum_index = HEADER_SIZE + packet_2_length - checksum_len;
-				let checksum = checksum::calculate_checksum(0, &_buffer[offset..offset + checksum_index]);
+				let checksum = crate::checksum::calculate_checksum(0, &_buffer[offset..offset + checksum_index]);
 				endian::write_u16_le(&mut _buffer[offset + checksum_index..], checksum);
 				Ok(50)
 			}
@@ -366,7 +290,7 @@ mod test {
 		}
 
 		// Setup the bus with a dummy serial interface
-		let mut bus = Bus::with_buffers(DummySerial {}, read_buffer, write_buffer).unwrap();
+		let mut bus = sync::Bus::with_buffers(DummySerial {}, read_buffer, write_buffer).unwrap();
 
 		// Read the corrupt package
 		let deadline = std::time::Instant::now() + Duration::from_secs(1);
