+++ src/client.rs
@@ -1,6 +1,8 @@
-use super::bus::Bus;
+use super::Bus;
+use super::Serial2Port;
+use super::SerialPort;
 use crate::instruction_id;
-use crate::{bus_types::StatusPacket, ReadError, TransferError, WriteError};
+use crate::{bus::StatusPacket, ReadError, TransferError, WriteError};
 use core::time::Duration;
 
 macro_rules! make_client_struct {
@@ -15,25 +17,25 @@ macro_rules! make_client_struct {
 		/// The `Buffer` generic type argument defaults to `Vec<u8>` if the `"alloc"` feature is enabled,
 		/// and to `&'static mut [u8]` otherwise.
 		/// See the [`crate::static_buffer!()`] macro for a way to safely create a mutable static buffer.
-		pub struct Client<SerialPort $(= $DefaultSerialPort)?, Buffer = crate::bus::DefaultBuffer>
+		pub struct Client<Port $(= $DefaultSerialPort)?, Buffer = crate::bus::DefaultBuffer>
 		where
-			SerialPort: super::SerialPort,
+			Port: SerialPort,
 			Buffer: AsRef<[u8]> + AsMut<[u8]>,
 		{
-			bus: Bus<SerialPort, Buffer>,
+			bus: Bus<Port, Buffer>,
 		}
 	};
 }
 
 #[cfg(feature = "serial2")]
-make_client_struct!(super::Serial2Port);
+make_client_struct!(Serial2Port);
 
 #[cfg(not(feature = "serial2"))]
 make_client_struct!();
 
-impl<SerialPort, Buffer> core::fmt::Debug for Client<SerialPort, Buffer>
+impl<Port, Buffer> core::fmt::Debug for Client<Port, Buffer>
 where
-	SerialPort: super::SerialPort + core::fmt::Debug,
+	Port: SerialPort + core::fmt::Debug,
 	Buffer: AsRef<[u8]> + AsMut<[u8]>,
 {
 	fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
@@ -45,20 +47,20 @@ where
 }
 
 #[cfg(feature = "serial2")]
-impl Client<super::Serial2Port, Vec<u8>> {
+impl Client<Serial2Port, Vec<u8>> {
 	/// Open a serial port with the given baud rate.
 	///
 	/// This will allocate a new read and write buffer of 128 bytes each.
 	/// Use [`Self::open_with_buffers()`] if you want to use a custom buffers.
 	pub fn open(path: impl AsRef<std::path::Path>, baud_rate: u32) -> std::io::Result<Self> {
-		let serial_port = super::Serial2Port::open(path, baud_rate)?;
+		let serial_port = Serial2Port::open(path, baud_rate)?;
 		let bus = Bus::with_buffers_and_baud_rate(serial_port, vec![0; 128], vec![0; 128], baud_rate);
 		Ok(Self { bus })
 	}
 }
 
 #[cfg(feature = "serial2")]
-impl<Buffer> Client<super::Serial2Port, Buffer>
+impl<Buffer> Client<Serial2Port, Buffer>
 where
 	Buffer: AsRef<[u8]> + AsMut<[u8]>,
 {
@@ -71,16 +73,16 @@ where
 		read_buffer: Buffer,
 		write_buffer: Buffer,
 	) -> std::io::Result<Self> {
-		let serial_port = super::Serial2Port::open(path, baud_rate)?;
+		let serial_port = Serial2Port::open(path, baud_rate)?;
 		let bus = Bus::with_buffers_and_baud_rate(serial_port, read_buffer, write_buffer, baud_rate);
 		Ok(Self { bus })
 	}
 }
 
 #[cfg(feature = "alloc")]
-impl<SerialPort> Client<SerialPort, alloc::vec::Vec<u8>>
+impl<Port> Client<Port, alloc::vec::Vec<u8>>
 where
-	SerialPort: super::SerialPort,
+	Port: SerialPort,
 {
 	/// Create a new client using an open serial port.
 	///
@@ -90,23 +92,23 @@ where
 	/// This will allocate a new read and write buffer of 128 bytes each.
 	/// Use [`Self::with_buffers()`] if you want to use a custom buffers.
 	#[cfg(feature = "alloc")]
-	pub fn new(serial_port: SerialPort) -> Result<Self, SerialPort::Error> {
+	pub fn new(serial_port: Port) -> Result<Self, Port::Error> {
 		let bus = Bus::with_buffers(serial_port, alloc::vec![0; 128], alloc::vec![0; 128])?;
 		Ok(Self { bus })
 	}
 }
 
 #[super::bisync]
-impl<SerialPort, Buffer> Client<SerialPort, Buffer>
+impl<Port, Buffer> Client<Port, Buffer>
 where
-	SerialPort: super::SerialPort,
+	Port: SerialPort,
 	Buffer: AsRef<[u8]> + AsMut<[u8]>,
 {
 	/// Create a new client using pre-allocated buffers.
 	///
 	/// The serial port must already be configured in raw mode with the correct baud rate,
 	/// character size (8), parity (disabled) and stop bits (1).
-	pub fn with_buffers(serial_port: SerialPort, read_buffer: Buffer, write_buffer: Buffer) -> Result<Self, SerialPort::Error> {
+	pub fn with_buffers(serial_port: Port, read_buffer: Buffer, write_buffer: Buffer) -> Result<Self, Port::Error> {
 		let bus = Bus::with_buffers(serial_port, read_buffer, write_buffer)?;
 		Ok(Self { bus })
 	}
@@ -117,7 +119,7 @@ where
 	/// and may disrupt the communication with the motors.
 	/// In general, it should be safe to read and write to the bus manually in between instructions,
 	/// if the response from the motors has already been received.
-	pub fn serial_port(&self) -> &SerialPort {
+	pub fn serial_port(&self) -> &Port {
 		&self.bus.serial_port
 	}
 
@@ -125,7 +127,7 @@ where
 	///
 	/// This discards any data in internal the read buffer of the client.
 	/// This is normally not a problem, since all data in the read buffer is also discarded when transmitting a new command.
-	pub fn into_serial_port(self) -> SerialPort {
+	pub fn into_serial_port(self) -> Port {
 		self.bus.serial_port
 	}
 
@@ -135,7 +137,7 @@ where
 	}
 
 	/// Set the baud rate of the underlying serial port.
-	pub fn set_baud_rate(&mut self, baud_rate: u32) -> Result<(), SerialPort::Error> {
+	pub fn set_baud_rate(&mut self, baud_rate: u32) -> Result<(), Port::Error> {
 		self.bus.set_baud_rate(baud_rate)
 	}
 
@@ -153,7 +155,7 @@ where
 		parameter_count: usize,
 		expected_response_parameters: u16,
 		encode_parameters: F,
-	) -> Result<StatusPacket<'_>, TransferError<SerialPort::Error>>
+	) -> Result<StatusPacket<'_>, TransferError<Port::Error>>
 	where
 		F: FnOnce(&mut [u8]) -> Result<(), crate::error::BufferTooSmallError>,
 	{
@@ -171,7 +173,7 @@ where
 		instruction_id: u8,
 		parameter_count: usize,
 		encode_parameters: F,
-	) -> Result<(), WriteError<SerialPort::Error>>
+	) -> Result<(), WriteError<Port::Error>>
 	where
 		F: FnOnce(&mut [u8]) -> Result<(), crate::error::BufferTooSmallError>,
 	{
@@ -181,10 +183,7 @@ where
 	}
 
 	/// Read a raw status response from the bus with the given deadline.
-	pub async fn read_status_response_timeout<'a>(
-		&'a mut self,
-		timeout: Duration,
-	) -> Result<StatusPacket<'a>, ReadError<SerialPort::Error>> {
+	pub async fn read_status_response_timeout<'a>(&'a mut self, timeout: Duration) -> Result<StatusPacket<'a>, ReadError<Port::Error>> {
 		let deadline = self.serial_port().make_deadline(timeout);
 		let packet = self.bus.read_packet_deadline(deadline).await?;
 		let status = match packet.as_status() {
@@ -205,13 +204,10 @@ where
 	/// Read a raw status response with an automatically calculated timeout.
 	///
 	/// The read timeout is determined by the expected number of response parameters and the baud rate of the bus.
-	pub async fn read_status_response<'a>(
-		&'a mut self,
-		expected_parameters: u16,
-	) -> Result<StatusPacket<'a>, ReadError<SerialPort::Error>> {
+	pub async fn read_status_response<'a>(&'a mut self, expected_parameters: u16) -> Result<StatusPacket<'a>, ReadError<Port::Error>> {
 		// Official SDK adds a flat 34 milliseconds, so lets just mimick that.
-		let message_size = crate::bus_types::StatusPacket::message_len(expected_parameters as usize) as u32;
-		let timeout = crate::bus_types::message_transfer_time(message_size, self.bus.baud_rate) + Duration::from_millis(34);
+		let message_size = crate::bus::StatusPacket::message_len(expected_parameters as usize) as u32;
+		let timeout = crate::bus::message_transfer_time(message_size, self.bus.baud_rate) + Duration::from_millis(34);
 		self.read_status_response_timeout(timeout).await
 	}
 }
